sw/robot-api: enable tests
This requires https://github.com/sourcebots/robot-api/pull/33 to be
merged; at the moment I'm building with:

    nix-build -E 'import ./default.nix; robot-api.overrideDerivation (_: { src = ../robot-api; })'

to override the package source to my local checkout. Once the PR is
merged I'll bump the source version so that this isn't necessary.

The tests are made optional since robotd and sb-vision are somewhat
hefty dependencies to pull in.

Note that the tests are currently failing.